+++ packages/ignition/src/app/components/opengraph.tsx
@@ -1,69 +1,19 @@
+import { readFileSync } from "fs";
+import { resolve } from "path";
 import { ImageResponse } from "next/og";
 import { Languages } from "@/types";
-import { serverEnv } from "@/env/server";
 import { withLocale } from "@/locales";
-import { FaFly, FaIcons, FaRoad } from "rocketicons/fa";
-import { CollectionID, IconsManifest } from "rocketicons/data";
-import { IconType } from "rocketicons";
+import { CollectionID } from "rocketicons/data";
+import { collectionsCounts, total } from "@/data-helpers/icons/manifest";
+import { IconTree, Variants } from "rocketicons";
 import { BiCollection } from "rocketicons/bi";
 import { TbIcons } from "rocketicons/tb";
-import { RcRocketIcon } from "rocketicons/rc";
-import { SlDocs } from "rocketicons/sl";
-import { LuBird, LuSmile } from "rocketicons/lu";
-import { PiAlien, PiFlyingSaucer } from "rocketicons/pi";
 import { BsCollection } from "rocketicons/bs";
+import React from "react";
+import { tree2Element } from "rocketicons/core/utils";
 
 const numberFormatter = (lang: Languages, number: number) =>
   new Intl.NumberFormat(lang).format(number);
-
-const selectRandomIcon = (
-  style: React.CSSProperties | undefined,
-  iconKey?: string
-): IconType | undefined => {
-  const iconsMap = new Map<string, any>([
-    ["LuSmile", <LuSmile key={"lusmile"} style={style} />],
-    ["LuBird", <LuBird key={"lubird"} style={style} />],
-    ["FaFly", <FaFly key={"fafly"} style={style} />],
-    ["PiFlyingSaucer", <PiFlyingSaucer key={"piflyingsaucer"} style={style} />],
-    ["PiAlien", <PiAlien key={"pialien"} style={style} />]
-  ]);
-
-  if (iconKey && iconsMap.has(iconKey)) {
-    return iconsMap.get(iconKey);
-  }
-  const iconsArray = Array.from(iconsMap.values());
-  const randomIndex = Math.floor(Math.random() * iconsArray.length);
-
-  return iconsArray[randomIndex];
-};
-
-export const RocketIconChooser = ({
-  subheading,
-  style,
-  iconKey,
-  Icon
-}: {
-  subheading?: string;
-  style?: React.CSSProperties;
-  iconKey?: string;
-  Icon?: IconType;
-}): any => {
-  if (Icon) {
-    return <Icon style={style} />;
-  } else if (subheading) {
-    if (subheading.startsWith("/docs")) {
-      return <SlDocs style={style} />;
-    } else if (subheading.startsWith("/icons")) {
-      return <FaIcons style={style} />;
-    } else if (subheading.startsWith("roadmap")) {
-      return <FaRoad style={style} />;
-    } else {
-      return selectRandomIcon(style, iconKey);
-    }
-  } else {
-    return <RcRocketIcon style={style} />;
-  }
-};
 
 const OpenGraph = async ({
   lang,
@@ -72,9 +22,9 @@ const OpenGraph = async ({
   iconCollectionCount,
   iconCollectionName,
   iconName,
+  iconJson,
   text,
-  darkMode = true,
-  Icon
+  darkMode = true
 }: {
   lang: Languages;
   subheading?: string;
@@ -82,70 +32,55 @@ const OpenGraph = async ({
   iconCollectionCount?: number;
   iconCollectionName?: string;
   iconName?: string;
+  iconJson?: { variant: Variants; iconTree: IconTree };
   text?: string;
   darkMode?: boolean;
-  Icon?: IconType;
 }) => {
-  const opengraph = withLocale(lang).config("opengraph");
-
-  const logoImg = darkMode
-    ? "logo-rocketicons-white-nobg-512.png"
-    : "logo-rocketicons-black-nobg-512.png";
   const bigIconSize = 200;
   const smallIconSize = 28;
   const color = darkMode ? "#ddd" : "#444";
+  const outerPaddingClass = "p-80px";
+  const internalLeftMarginClass = "ml-20px";
+
   const textColor = color,
     iconColor = color;
   const textGradient = `linear-gradient(to bottom right, ${
     darkMode ? "#fff 20%, #0ea5e9 70%" : "#000000 20%, #0ea5e9 70%"
   })`;
 
-  const outerPaddingClass = "p-80px";
-  const internalLeftMarginClass = "ml-20px";
+  const opengraph = withLocale(lang).config("opengraph");
 
-  const bgImage = darkMode ? "og-hero-dark.jpg" : "og-hero-light.jpg";
+  const quicksandRegular = readFileSync(resolve("./public", "fonts", "Quicksand-Regular.ttf"));
+  const interMedium = readFileSync(resolve("./public", "fonts", "Inter-Medium.ttf"));
 
-  const quicksandRegular = fetch(
-    new URL(`${serverEnv.NEXT_PUBLIC_APP_URL}/fonts/Quicksand-Regular.ttf`, import.meta.url)
-  ).then((res) => res.arrayBuffer());
-
-  const interMedium = fetch(
-    new URL(`${serverEnv.NEXT_PUBLIC_APP_URL}/fonts/Inter-Medium.ttf`, import.meta.url)
-  ).then((res) => res.arrayBuffer());
-
-  const firaCodeRegular = fetch(
-    new URL(`${serverEnv.NEXT_PUBLIC_APP_URL}/fonts/FiraCode-Regular.ttf`, import.meta.url)
-  ).then((res) => res.arrayBuffer());
-
-  const groupedCollections: Map<string, number> = new Map();
-  IconsManifest.forEach(({ id, icons }: { id: string; icons: any[] }) => {
-    if (!groupedCollections.get(id)) {
-      groupedCollections.set(id, 0);
-    }
-    groupedCollections.set(id, icons.length);
-  });
-
-  const totalIconsCount = Array.from(groupedCollections.values()).reduce(
-    (acc, count) => acc + count,
-    0
+  const logoImg = readFileSync(
+    resolve(
+      "./public",
+      darkMode ? "logo-rocketicons-white-nobg-512.png" : "logo-rocketicons-black-nobg-512.png"
+    ),
+    { encoding: "base64" }
   );
 
+  const totalIconsCount = total;
+
   const brand = withLocale(lang).config("brand");
+
+  const bgImg = readFileSync(
+    resolve("./public", darkMode ? "img/og-hero-dark.jpg" : "img/og-hero-light.jpg"),
+    { encoding: "base64" }
+  );
 
   const mainDivStyle = {
     display: "flex",
     color: textColor,
-    backgroundImage: `url("${serverEnv.NEXT_PUBLIC_APP_URL}/img/${bgImage}")`,
+    backgroundImage: `url("data:image/jpeg;base64,${bgImg}")`,
     backgroundSize: "1200px 630px"
   };
 
   const bigIconsStyle = {
     width: `${bigIconSize}px`,
     height: `${bigIconSize}px`,
-    display: "block",
-    fill: iconColor,
-    stroke: iconColor,
-    color: iconColor,
+    display: "flex",
     padding: "0px"
   };
 
@@ -184,7 +119,7 @@ const OpenGraph = async ({
               <div tw="flex flex-row">
                 <picture>
                   <img
-                    src={`${serverEnv.NEXT_PUBLIC_APP_URL}/${logoImg}`}
+                    src={`data:image/png;base64,${logoImg}`}
                     alt="rocketicons Logo"
                     tw="w-128 h-23"
                     width={128}
@@ -203,9 +138,20 @@ const OpenGraph = async ({
                 </p>
               )}
             </div>
-            <div tw="flex">
-              <RocketIconChooser subheading={subheading} style={bigIconsStyle} Icon={Icon} />
-            </div>
+            {iconJson && (
+              <div tw="flex">
+                <svg
+                  {...iconJson.iconTree.attr}
+                  style={{
+                    ...bigIconsStyle,
+                    ...(["filled", "full"].includes(iconJson.variant) && { fill: iconColor }),
+                    ...(["outlined", "full"].includes(iconJson.variant) && { stroke: iconColor })
+                  }}
+                >
+                  {tree2Element(iconJson.iconTree.child)}
+                </svg>
+              </div>
+            )}
           </div>
         </div>
         <div tw={`flex grow mt-10 ${internalLeftMarginClass}`}>
@@ -229,7 +175,7 @@ const OpenGraph = async ({
           <div tw="flex flex-row grow">
             <BiCollection style={smallIconStyle} />
             <div tw="flex flex-col ml-3">
-              <span>{numberFormatter(lang, groupedCollections.size)}</span>
+              <span>{numberFormatter(lang, collectionsCounts.size)}</span>
               <span>{opengraph["collections"]}</span>
             </div>
           </div>
@@ -249,17 +195,12 @@ const OpenGraph = async ({
       fonts: [
         {
           name: "Inter",
-          data: await interMedium,
+          data: interMedium,
           style: "normal"
         },
         {
           name: "Quicksand",
-          data: await quicksandRegular,
-          style: "normal"
-        },
-        {
-          name: "FiraCode",
-          data: await firaCodeRegular,
+          data: quicksandRegular,
           style: "normal"
         }
       ]