+++ packages/ignition/src/statics-generator/collections-static-params.ts
@@ -3,7 +3,7 @@ import { getManifest, write } from "./utils";
 import { Languages, AvailableLanguages } from "@/types";
 import { serverEnv } from "@/env/server";
 
-const OUTPUT_FILE = "params/icons.json";
+const OUTPUT_FILE = "params/collections.json";
 
 const generator = async () => {
   const params = serverEnv.SKIP_BUILD_STATIC_GENERATION